fix(lightspeed): moved CustomDrawer out of the packages/app
