+++ workspaces/lightspeed/.changeset/soft-yaks-hope.md
@@ -1,0 +1,5 @@
+---
+'@red-hat-developer-hub/backstage-plugin-lightspeed': patch
+---
+
+Moved the `CustomDrawer` component from `packages/app` to `plugins/lightspeed/dev` to resolve workspace accessibility issues during development.
