+++ workspaces/lightspeed/packages/app/package.json
@@ -47,7 +47,7 @@
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@mui/material": "^5.12.2",
-    "@red-hat-developer-hub/backstage-plugin-lightspeed": "*",
+    "@red-hat-developer-hub/backstage-plugin-lightspeed": "workspace:^",
     "@red-hat-developer-hub/backstage-plugin-theme": "^0.12.0",
     "react": "^18.0.2",
     "react-dom": "^18.0.2",
