+++ workspaces/lightspeed/packages/app/src/components/Root/ApplicationDrawer.tsx
@@ -22,7 +22,7 @@ import {
   useEffect,
   useRef,
 } from 'react';
-import { CustomDrawer } from './CustomDrawer';
+import { CustomDrawer } from '@red-hat-developer-hub/backstage-plugin-lightspeed/dev';
 
 /**
  * Partial drawer state exposed by drawer plugins
