+++ workspaces/lightspeed/plugins/lightspeed/dev/DrawerComponent.tsx
@@ -18,9 +18,8 @@ import { PropsWithChildren } from 'react';
 
 import { ChatbotDisplayMode } from '@patternfly/chatbot';
 
-// eslint-disable-next-line @backstage/no-relative-monorepo-imports
-import { CustomDrawer } from '../../../packages/app/src/components/Root/CustomDrawer';
 import { useLightspeedDrawerContext } from '../src/hooks/useLightspeedDrawerContext';
+import { CustomDrawer } from './CustomDrawer';
 
 export const DrawerComponent = ({ children }: PropsWithChildren) => {
   const { displayMode, drawerWidth, setDrawerWidth } =
