+++ workspaces/lightspeed/plugins/lightspeed/package.json
@@ -10,12 +10,16 @@
   "exports": {
     ".": "./src/index.ts",
     "./alpha": "./src/alpha.ts",
+    "./dev": "./dev/CustomDrawer.tsx",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "alpha": [
         "src/alpha.ts"
+      ],
+      "dev": [
+        "dev/CustomDrawer.tsx"
       ],
       "package.json": [
         "package.json"
