+++ workspaces/lightspeed/plugins/lightspeed/report-dev.api.md
@@ -1,0 +1,7 @@
+## API Report File for "@red-hat-developer-hub/backstage-plugin-lightspeed"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+// (No @packageDocumentation comment for this package)
+```
